requirements.txt
python-socketio==5.11.4
python-engineio==4.10.1
pyaudio==0.2.13
wave==0.0.2
webrtcvad==2.0.10
loguru==0.7.0
requests==2.31.0
retrying==1.3.4
pydub==0.25.1
schedule==1.1.0
SpeechRecognition==3.8.1
vosk==0.3.45
ujson==5.8.0
psutil==5.9.5
python-decouple==3.8
